+++ huaweicloud-sdk-java-dis/src/main/java/com/huaweicloud/dis/DISClientAsync2.java
@@ -2,6 +2,10 @@ package com.huaweicloud.dis;
 
 import java.util.concurrent.Future;
 
+import com.huaweicloud.dis.iface.app.request.ListStreamConsumingStateRequest;
+import com.huaweicloud.dis.iface.app.response.ListStreamConsumingStateResult;
+import com.huaweicloud.dis.iface.data.request.*;
+import com.huaweicloud.dis.iface.data.response.*;
 import org.apache.http.HttpRequest;
 
 import com.huaweicloud.dis.DISConfig.BodySerializeType;
@@ -21,18 +25,6 @@ import com.huaweicloud.dis.iface.app.request.CreateAppRequest;
 import com.huaweicloud.dis.iface.app.request.ListAppsRequest;
 import com.huaweicloud.dis.iface.app.response.DescribeAppResult;
 import com.huaweicloud.dis.iface.app.response.ListAppsResult;
-import com.huaweicloud.dis.iface.data.request.CommitCheckpointRequest;
-import com.huaweicloud.dis.iface.data.request.GetCheckpointRequest;
-import com.huaweicloud.dis.iface.data.request.GetPartitionCursorRequest;
-import com.huaweicloud.dis.iface.data.request.GetRecordsRequest;
-import com.huaweicloud.dis.iface.data.request.PutFilesRequest;
-import com.huaweicloud.dis.iface.data.request.PutRecordsRequest;
-import com.huaweicloud.dis.iface.data.response.CommitCheckpointResult;
-import com.huaweicloud.dis.iface.data.response.GetCheckpointResult;
-import com.huaweicloud.dis.iface.data.response.GetPartitionCursorResult;
-import com.huaweicloud.dis.iface.data.response.GetRecordsResult;
-import com.huaweicloud.dis.iface.data.response.PutFilesResult;
-import com.huaweicloud.dis.iface.data.response.PutRecordsResult;
 import com.huaweicloud.dis.iface.stream.request.CreateStreamRequest;
 import com.huaweicloud.dis.iface.stream.request.DeleteStreamRequest;
 import com.huaweicloud.dis.iface.stream.request.DescribeStreamRequest;
@@ -221,6 +213,37 @@ public class DISClientAsync2 extends AbstractDISClientAsync implements DISAsync{
                 .build());
     	
         return requestAsync(getPartitionCursorParam, request, GetPartitionCursorResult.class, asyncHandler);
+	}
+
+	@Override
+	public Future<DeleteCheckpointResult> deleteCheckpointAsync(DeleteCheckpointRequest deleteCheckpointRequest) {
+		return deleteCheckpointAsync(deleteCheckpointRequest,null);
+	}
+
+	@Override
+	public Future<DeleteCheckpointResult> deleteCheckpointAsync(DeleteCheckpointRequest deleteCheckpointRequest, AsyncHandler<DeleteCheckpointResult> asyncHandler) {
+		Request<HttpRequest> request = buildRequest(HttpMethodName.DELETE, disConfig.getEndpoint(),
+				ResourcePathBuilder.standard()
+						.withProjectId(disConfig.getProjectId())
+						.withResource(new CheckPointResource(null))
+						.build());
+		return requestAsync(deleteCheckpointRequest, request, DeleteCheckpointResult.class, asyncHandler);
+	}
+
+	@Override
+	public Future<ListStreamConsumingStateResult> listStreamConsumingStateAsync(ListStreamConsumingStateRequest listStreamConsumingStateRequest) {
+		return listStreamConsumingStateAsync(listStreamConsumingStateRequest,null);
+	}
+
+	@Override
+	public Future<ListStreamConsumingStateResult> listStreamConsumingStateAsync(ListStreamConsumingStateRequest listStreamConsumingStateRequest, AsyncHandler<ListStreamConsumingStateResult> asyncHandler) {
+		Request<HttpRequest> request = buildRequest(HttpMethodName.GET, disConfig.getEndpoint(),
+				ResourcePathBuilder.standard()
+						.withProjectId(disConfig.getProjectId())
+						.withResource(new AppsResource(listStreamConsumingStateRequest.getAppName()))
+						.withResource(new StreamResource(listStreamConsumingStateRequest.getStreamName()))
+						.build());
+		return requestAsync(listStreamConsumingStateRequest, request, ListStreamConsumingStateResult.class, asyncHandler);
 	}
 
 	@Override