+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 
 - Features
   * add describeApp and listApps features
+  * add delete checkpoint
+  * add describe consumer state